Refactor `strategy_templates.py` into modules for each strategy
