+++ tests/test_backtester.py
@@ -8,11 +8,15 @@ from typing import Any
 import polars as pl
 import pytest
 from quant_trading_strategy_backtester.backtester import Backtester
-from quant_trading_strategy_backtester.strategy_templates import (
+from quant_trading_strategy_backtester.strategies.base import Strategy
+from quant_trading_strategy_backtester.strategies.mean_reversion import (
     MeanReversionStrategy,
+)
+from quant_trading_strategy_backtester.strategies.moving_average_crossover import (
     MovingAverageCrossoverStrategy,
+)
+from quant_trading_strategy_backtester.strategies.pairs_trading import (
     PairsTradingStrategy,
-    Strategy,
 )
